Mise à jour commentaires

+++ src/main/java/com/lambazon/parkingsystem/service/FareCalculatorService.java
@@ -5,7 +5,7 @@ import com.lambazon.parkingsystem.model.Ticket;
 
 public class FareCalculatorService {
 
-    // Default : no discount
+    // Default : Pas de remise
     public void calculateFare(Ticket ticket) {
         calculateFare(ticket, false);
     }
@@ -15,15 +15,13 @@ public class FareCalculatorService {
             throw new IllegalArgumentException("Out time provided is incorrect:"+ticket.getOutTime().toString());
         }
 
-        //Conversion en millisecondes
-        long inHour = ticket.getInTime().getTime();
-        long outHour = ticket.getOutTime().getTime();
+        double inHour = ticket.getInTime().getTime();
+        double outHour = ticket.getOutTime().getTime();
 
-        //TODO: Some tests are failing here. Need to check if this logic is correct
-        //Conversion en heure et utilisation des double
-        double duration = (double) (outHour - inHour) /(60000*60);
+        //Conversion en heure
+        double duration = (outHour - inHour) /(60000*60);
 
-        //US1 : Free parking for 30min
+        //Parking gratuit pour 30min
         if (duration <= 0.5){
             duration = 0;
         }
@@ -39,7 +37,7 @@ public class FareCalculatorService {
             }
             default: throw new IllegalArgumentException("Unkown Parking Type");
         }
-        // US 2 : Apply discount if the discount flag is true
+        //Remise si discount est true
         if (discount) {
             ticket.setPrice(ticket.getPrice() * 0.95);
         }
